// frontend/src/app/models/izvestaj.ts
export class Izvestaj{
    pacijent: string;
    lekar: string;
    ime_lekara: string;
    prezime_lekara: string;
    specijalizacija: string;
    razlog_dolaska: string;
    dijagnoza: string;
    terapija: string;
    datum_pregleda: string;
    datum_narednog_pregleda: string;
    formatiran_datum: string;
    formatirano_vreme: string;
    formatiran_datum_naredni: string;
}